backend/requirements.txt
asgiref==3.6.0
boto3==1.26.79
botocore==1.29.79
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
dropbox==11.36.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
ply==3.11
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
uritemplate==4.1.1
urllib3==1.26.14
git+https://github.com/getcohesive/marketplace-server-plugin-django.git@usage_tracking#egg=cohesive-marketplace-server-plugin-django
